Move ``get_world_size_and_rank`` to utils (#2155)
