+++ tests/torchtune/utils/test_device.py
@@ -12,6 +12,8 @@ from unittest.mock import patch
 import pytest
 
 import torch
+
+from torch.distributed import launcher
 from torchtune.utils._device import (
     _get_device_type_from_env,
     _setup_device,
@@ -20,12 +22,31 @@ from torchtune.utils._device import (
     get_device,
     get_device_support,
     get_torch_device_namespace,
+    get_world_size_and_rank,
 )
 
 
 class TestDevice:
 
     cuda_available: bool = torch.cuda.is_available()
+
+    def _create_world(self, expected_world_size: int) -> None:
+        torch.distributed.init_process_group(backend="gloo")
+        world_size, _ = get_world_size_and_rank()
+        if world_size != expected_world_size:
+            raise AssertionError(
+                f"Expected different world size: received {world_size}, expected {expected_world_size}"
+            )
+
+    def test_world_size_with_cpu(self, get_pet_launch_config) -> None:
+        desired_world_size = 4
+        lc = get_pet_launch_config(desired_world_size)
+        launcher.elastic_launch(lc, entrypoint=self._create_world)(desired_world_size)
+
+    def test_rank_with_cpu_device(self) -> None:
+        """Very, very basic test"""
+        _, rank = get_world_size_and_rank()
+        assert rank == 0
 
     @patch("torch.cuda.is_available", return_value=False)
     def test_get_cpu_device(self, mock_cuda):
