+++ torchtune/training/metric_logging.py
@@ -14,9 +14,8 @@ import torch
 
 from numpy import ndarray
 from omegaconf import DictConfig, OmegaConf
-from torchtune.training._distributed import get_world_size_and_rank
 
-from torchtune.utils import get_logger
+from torchtune.utils import get_logger, get_world_size_and_rank
 from typing_extensions import Protocol
 
 Scalar = Union[torch.Tensor, ndarray, int, float]
