+++ torchtune/training/seed.py
@@ -13,8 +13,8 @@ from typing import Optional, Union
 import numpy as np
 import torch
 
-from torchtune.training._distributed import _broadcast_tensor, get_world_size_and_rank
-from torchtune.utils import get_logger
+from torchtune.training._distributed import _broadcast_tensor
+from torchtune.utils import get_logger, get_world_size_and_rank
 
 _log: logging.Logger = get_logger()
